Drinking state

+++ Assets/Model/Animals/State/Resting.cs
@@ -14,7 +14,7 @@ namespace Safari.Model.Animals.State
 
         private float restingDuration;
 
-        public Resting(Animal owner, int thirst) : base(owner, thirst)
+        public Resting(Animal owner, float thirst) : base(owner, thirst)
         {
         }
